backend/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.1.1
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.0
blinker==1.6.3
CacheControl==0.13.1
cachetools==5.3.2
celery==5.3.5
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.5
dnspython==2.4.2
firebase-admin==6.2.0
Flask==3.0.0
Flask-Cors==4.0.0
flower==2.0.1
frozenlist==1.4.0
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.0
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
httplib2==0.22.0
humanize==4.9.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.3
MarkupSafe==2.1.3
msgpack==1.0.7
multidict==6.0.4
openai==0.28.1
prometheus-client==0.19.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pymongo==4.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
vine==5.1.0
wcwidth==0.2.9
Werkzeug==3.0.0
yarl==1.9.2
zipp==3.17.0
